+++ love/src/components/MainTel/M1M3/M1M3.container.jsx
@@ -26,6 +26,7 @@ import {
   getM1M3ActuatorForces,
   getM1M3HardpointMonitorData,
   getM1M3HardpointActuatorState,
+  getAlignmentState,
 } from 'redux/selectors';
 import SubscriptionTableContainer from 'components/GeneralPurpose/SubscriptionTable/SubscriptionTable.container';
 import M1M3 from './M1M3';
@@ -63,7 +64,15 @@ const mapStateToProps = (state) => {
   const actuatorsForces = getM1M3ActuatorForces(state);
   const hardpointsMonitor = getM1M3HardpointMonitorData(state);
   const hardpointsActuatorState = getM1M3HardpointActuatorState(state);
-  return { ...m1m3State, ...actuatorsState, ...actuatorsForces, ...hardpointsMonitor, ...hardpointsActuatorState };
+  const alignmentState = getAlignmentState(state);
+  return {
+    ...m1m3State,
+    ...actuatorsState,
+    ...actuatorsForces,
+    ...hardpointsMonitor,
+    ...hardpointsActuatorState,
+    ...alignmentState,
+  };
 };
 
 const mapDispatchToProps = (dispatch) => {
@@ -90,6 +99,7 @@ const mapDispatchToProps = (dispatch) => {
     'event-MTM1M3-0-appliedOffsetForces',
     'event-MTM1M3-0-appliedStaticForces',
     'event-MTM1M3-0-enabledForceActuators',
+    'event-OAA-0-alignment',
     // 'event-MTM1M3-0-preclippedAccelerationForces',
     // 'event-MTM1M3-0-preclippedActiveOpticForces',
     // 'event-MTM1M3-0-preclippedBalanceForces',